Moved login and signup to user folder
