+++ app/javascript/components/user/Login.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useDispatch } from "react-redux";
 import { Link, useNavigate } from "react-router-dom";
-import { loginUser } from "../redux/userslice";
+import { loginUser } from "../../redux/userslice";
 
 const Login = () => {
   const [formData, setFormData] = useState({ email: "", password: "" });
